Format code with ESLint
