+++ frontend/src/App.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
-import NavBar from './components/navbar.js';
-import HomePage from './pages/home.js';
-import SearchPage from './pages/search.js';
-import AboutPage from './pages/about.js';
-import SignupPage from './pages/signup.js';
+import NavBar from './components/navbar';
+import HomePage from './pages/home';
+import SearchPage from './pages/search';
+import AboutPage from './pages/about';
+import SignupPage from './pages/signup';
 import SignupVolunteerPage from './pages/signup_volunteer';
-import StoryBook from './pages/storybook.js';
-import NewProject from './pages/newproject.js';
-import Style from './pages/style.js';
+import StoryBook from './pages/storybook';
+import NewProject from './pages/newproject';
+import Style from './pages/style';
 import Project from './pages/project';
 import UserProfile from './components/UserProfile/Profile.component';
 import Profile from './components/Profile/Profile.component';
