+++ frontend/src/components/Profile/Profile.component.jsx
@@ -11,21 +11,24 @@ import ProjectTitle from './Project-title.component';
 import ProjectName from './Project-name.component';
 
 class Profile extends Component {
-  state = {
-    logo: './src/img/crown-source-logo.svg',
-    profileName: 'John Doe',
-    role: 'Frontend Developer',
-    avatarURL: 'https://material-ui.com/static/images/avatar/1.jpg',
-    description: 'Contributers To 3 Active Projects',
-    state: 'New York',
-    country: 'USA',
-    descriptionTxt:
-      'Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dolorum, in! Culpa sequi, ipsum, cupiditate illum laboriosam, exercitationem odit adipisci numquam itaque atque animi. Soluta architecto eos ab necessitatibus aperiam voluptatibus. Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dolorum, in! Culpa sequi, ipsum, cupiditate illum laboriosam, exercitationem odit adipisci numquam itaque atque animi. Soluta architecto eos ab necessitatibus aperiam voluptatibus.',
-    labelIcon: <LabelIcon />,
-    label: 'Other Label',
-    projectImg:
-      'https://material-ui.com/static/images/cards/contemplative-reptile.jpg'
-  };
+  constructor(props) {
+    super(props);
+    this.state = {
+      logo: './src/img/crown-source-logo.svg',
+      profileName: 'John Doe',
+      role: 'Frontend Developer',
+      avatarURL: 'https://material-ui.com/static/images/avatar/1.jpg',
+      description: 'Contributers To 3 Active Projects',
+      state: 'New York',
+      country: 'USA',
+      descriptionTxt:
+        'Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dolorum, in! Culpa sequi, ipsum, cupiditate illum laboriosam, exercitationem odit adipisci numquam itaque atque animi. Soluta architecto eos ab necessitatibus aperiam voluptatibus. Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dolorum, in! Culpa sequi, ipsum, cupiditate illum laboriosam, exercitationem odit adipisci numquam itaque atque animi. Soluta architecto eos ab necessitatibus aperiam voluptatibus.',
+      labelIcon: <LabelIcon />,
+      label: 'Other Label',
+      projectImg:
+        'https://material-ui.com/static/images/cards/contemplative-reptile.jpg',
+    };
+  }
 
   render() {
     const {
@@ -39,10 +42,10 @@ class Profile extends Component {
       descriptionTxt,
       labelIcon,
       label,
-      projectImg
+      projectImg,
     } = this.state;
     return (
-      <Fragment>
+      <>
         <Header accountImg={avatarURL} logo={logo} />
         <AccountProfile
           accountImg={avatarURL}
@@ -55,7 +58,7 @@ class Profile extends Component {
         <Body txt={descriptionTxt} icon={labelIcon} label={label} />
         <ProjectTitle />
         <ProjectName img={projectImg} />
-      </Fragment>
+      </>
     );
   }
 }
