+++ frontend/src/constants/navigation.js
@@ -1,22 +1,22 @@
 export const HOME_PAGE_LINK = {
-	name: 'Home',
-	path: '/',
+  name: 'Home',
+  path: '/',
 };
 export const ABOUT_PAGE_LINK = {
-	name: 'About us',
-	path: '/about',
+  name: 'About us',
+  path: '/about',
 };
 export const SEARCH_PAGE_LINK = {
-	name: 'Search',
-	path: '/search',
+  name: 'Search',
+  path: '/search',
 };
 export const LOGIN_PAGE_LINK = {
-	name: 'Login',
-	path: '/login',
+  name: 'Login',
+  path: '/login',
 };
 export const SIGNUP_PAGE_LINK = {
-	name: 'Signup',
-	path: '/signup',
+  name: 'Signup',
+  path: '/signup',
 };
 
 export const PAGE_LINKS_LIST = [HOME_PAGE_LINK, ABOUT_PAGE_LINK, SEARCH_PAGE_LINK];
