+++ frontend/src/helpers/api.js
@@ -1,4 +1,5 @@
-import { getEnvVars } from '../environment';
+import getEnvVars from '../environment';
+
 const { BASE_URL } = getEnvVars();
 
 // FUTURE API - https://github.com/crowdsourcemedical/volunteer-database/wiki/Backend-API-spec
@@ -6,65 +7,65 @@ const { BASE_URL } = getEnvVars();
 // CURRENT API - http://localhost:8000/redoc
 
 class Api {
-    client = null;
+  async request({
+    route, payload, action = 'GET', headers = {},
+  }) {
+    const requestURL = `${BASE_URL}${route}`;
 
-    async request({ route, payload, action = 'GET', headers = {} }) {
-        const requestURL = `${BASE_URL}${route}`
+    const requestHeaders = {
+      'Content-Type': 'application/json',
+      ...headers,
+    };
 
-        const requestHeaders = {
-            'Content-Type': 'application/json',
-            ...headers
-        };
+    const options = {
+      method: action,
+      headers: requestHeaders,
+    };
 
-        const options = {
-            method: action,
-            headers: requestHeaders
-        }
-
-        if (payload) {
-            options.body = JSON.stringify(payload);
-        }
-
-        return fetch(requestURL, options).then(res => res.json()).catch(e => e);
+    if (payload) {
+      options.body = JSON.stringify(payload);
     }
 
-    async login(payload) {
-        const response = await this.request({
-            route: '/users/login',
-            action: 'POST',
-            payload,
-        })
+    return this.fetch(requestURL, options).then((res) => res.json()).catch((e) => e);
+  }
 
-        return response;
-    }
+  async login(payload) {
+    const response = await this.request({
+      route: '/users/login',
+      action: 'POST',
+      payload,
+    });
 
-    async getUsers() {
-        const response = await this.request({
-            route: '/users',
-            action: 'GET'
-        })
+    return response;
+  }
 
-        return response;  
-    }
+  async getUsers() {
+    const response = await this.request({
+      route: '/users',
+      action: 'GET',
+    });
 
-    async getUser(id) {
-        const response = await this.request({
-            route: `/users/${id}`,
-            action: 'GET'
-        })
+    return response;
+  }
 
-        return response;  
-    }
+  async getUser(id) {
+    const response = await this.request({
+      route: `/users/${id}`,
+      action: 'GET',
+    });
 
-    async createUser(payload) {
-        const response = await this.request({
-            route: `/users`,
-            action: 'POST',
-            payload
-        })
+    return response;
+  }
 
-        return response;  
-    }
+  async createUser(payload) {
+    const response = await this.request({
+      route: '/users',
+      action: 'POST',
+      payload,
+    });
+
+    return response;
+  }
 }
 
 export default new Api();
