+++ frontend/src/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { ThemeProvider } from '@material-ui/core/styles';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
 import theme from './styles/theme';
 
 ReactDOM.render(
