+++ frontend/src/pages/newproject.jsx
@@ -1,65 +1,67 @@
-import React, { useState } from "react";
-import TextField from "@material-ui/core/TextField";
-import { makeStyles } from "@material-ui/core/styles";
-import { Box, Typography, Grid } from "@material-ui/core";
-import { Button } from "@material-ui/core";
-import CheckBoxSection from "../components/Forms/checkboxSection.js";
+import React, { useState } from 'react';
+import TextField from '@material-ui/core/TextField';
+import { makeStyles } from '@material-ui/core/styles';
+import {
+  Box, Typography, Grid, Button,
+} from '@material-ui/core';
 
-const useStyles = makeStyles(theme => ({
+import CheckBoxSection from '../components/Forms/checkboxSection';
+
+const useStyles = makeStyles((theme) => ({
   root: {
-    "& > *": {
+    '& > *': {
       margin: theme.spacing(),
-      width: "25ch"
-    }
+      width: '25ch',
+    },
   },
   margin: {
-    margin: theme.spacing(1)
-  }
+    margin: theme.spacing(1),
+  },
 }));
 
 const checkboxesCad = [
-  { label: "Solid Works", status: false },
-  { label: "Inventor", status: false }
+  { label: 'Solid Works', status: false },
+  { label: 'Inventor', status: false },
 ]; // pass this to a checkboxsection to output all checkbox labels -> status is if its disabled or enabled
 const checkboxesMedical = [
-  { label: "Registered Nurse", status: false },
-  { label: "Physician Assistant", status: false },
-  { label: "Medical Student", status: false },
-  { label: "Intern", status: true },
-  { label: "Resident", status: false },
-  { label: "Attending", status: true }
+  { label: 'Registered Nurse', status: false },
+  { label: 'Physician Assistant', status: false },
+  { label: 'Medical Student', status: false },
+  { label: 'Intern', status: true },
+  { label: 'Resident', status: false },
+  { label: 'Attending', status: true },
 ]; // pass this to a checkboxsection to output all checkbox labels -> status is if its disabled or enabled
 const checkboxesEng = [
-  { label: "Mechanical Engineer", status: false },
-  { label: "Mechanical Engineer with FEA experience", status: false },
-  { label: "Electrical Engineer", status: true },
-  { label: "Mechatronics Engineer", status: false }
+  { label: 'Mechanical Engineer', status: false },
+  { label: 'Mechanical Engineer with FEA experience', status: false },
+  { label: 'Electrical Engineer', status: true },
+  { label: 'Mechatronics Engineer', status: false },
 ]; // pass this to a checkboxsection to output all checkbox labels -> status is if its disabled or enabled
 const checkboxesLegal = [
-  { label: "Lawyer", status: false },
-  { label: "Barrister", status: false },
-  { label: "paralegal", status: false }
+  { label: 'Lawyer', status: false },
+  { label: 'Barrister', status: false },
+  { label: 'paralegal', status: false },
 ];
 const checkboxesManu = [
-  { label: "FDM 3D printer", status: false },
-  { label: "SLA 3D printer", status: false },
-  { label: "SLS Nylon 3D printer", status: false },
-  { label: "Machinist", status: true }
+  { label: 'FDM 3D printer', status: false },
+  { label: 'SLA 3D printer', status: false },
+  { label: 'SLS Nylon 3D printer', status: false },
+  { label: 'Machinist', status: true },
 ];
-//TODO: populate the above with an API call to see which professions are avaliable and not avaliable
+// TODO: populate the above with an API call to see which professions are avaliable and not avaliable
 
 function NewProject() {
-  const [description, setDescription] = useState("");
-  const [profsRequired, setProfsRequired] = useState([]);
-  const [location, setLocation] = useState("");
+  const [description, setDescription] = useState('');
+  const [, setLocation] = useState('');
 
   const changeMap = {
     description: setDescription,
-    location: setLocation
+    location: setLocation,
   };
 
   function handleCheckBoxChange(event) {
-    //TODO: GET CHECKBOX data from child component onChange
+    global.console.log(event);
+    // TODO: GET CHECKBOX data from child component onChange
   }
 
   function handleChange(event) {
@@ -86,8 +88,9 @@ function NewProject() {
           <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
             <Typography color="primary" variant="h4">
               Tell us about your project.
-            </Typography>{" "}
-            <br></br>
+            </Typography>
+            {' '}
+            <br />
           </Grid>
           <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
             <Typography variant="h6">
@@ -98,9 +101,9 @@ function NewProject() {
           <form className={useStyles.root} noValidate autoComplete="off">
             <TextField
               name="description"
-              multiline={true}
+              multiline
               rows={10}
-              fullWidth={true}
+              fullWidth
               size="large"
               label="Description"
               variant="outlined"
@@ -114,8 +117,9 @@ function NewProject() {
                 color="primary"
                 variant="subtitle1"
               >
-                {" "}
-                Please read the recomendatiosn from our Legal{" "}
+                {' '}
+                Please read the recomendatiosn from our Legal
+                {' '}
               </Typography>
               <Box m={3} />
             </Grid>
@@ -137,45 +141,45 @@ function NewProject() {
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
               <CheckBoxSection
                 right
-                sectionHeader={"Medical Staff advisers"}
+                sectionHeader="Medical Staff advisers"
                 checkboxes={checkboxesMedical}
                 cb={handleCheckBoxChange}
-              ></CheckBoxSection>
+              />
             </Grid>
             <Box m={5} />
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
               <CheckBoxSection
                 right
-                sectionHeader={"Engineering"}
+                sectionHeader="Engineering"
                 checkboxes={checkboxesEng}
                 cb={handleCheckBoxChange}
-              ></CheckBoxSection>
+              />
             </Grid>
             <Box m={5} />
 
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
               <CheckBoxSection
                 right
-                sectionHeader={"CAD professionals"}
+                sectionHeader="CAD professionals"
                 checkboxes={checkboxesCad}
                 cb={handleCheckBoxChange}
-              ></CheckBoxSection>
+              />
             </Grid>
             <Box m={2} />
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
               <CheckBoxSection
-                sectionHeader={"Legal advisors"}
+                sectionHeader="Legal advisors"
                 checkboxes={checkboxesLegal}
                 cb={handleCheckBoxChange}
-              ></CheckBoxSection>
+              />
             </Grid>
             <Box m={2} />
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
               <CheckBoxSection
-                sectionHeader={"Manufactures"}
+                sectionHeader="Manufactures"
                 checkboxes={checkboxesManu}
                 cb={handleCheckBoxChange}
-              ></CheckBoxSection>
+              />
             </Grid>
             <Box m={3} />
             <Grid container alignItems="flex-start" item xs={12} lg={8} xl={8}>
