+++ frontend/src/pages/project.jsx
@@ -1,48 +1,50 @@
 import React from 'react';
-import { Grid, Typography, Divider, Hidden } from '@material-ui/core';
-import { MapPin } from 'react-feather';
-import { Bookmark } from 'react-feather';
-import { makeStyles } from '@material-ui/core/styles';
-import { Button } from '@material-ui/core';
+import {
+  Grid, Typography, Divider, Hidden, Button,
+} from '@material-ui/core';
+import { MapPin, Bookmark } from 'react-feather';
 
-const useStyles = makeStyles(theme => ({
+import { makeStyles } from '@material-ui/core/styles';
+
+
+const useStyles = makeStyles((theme) => ({
   root: {
     padding: theme.spacing(3),
     margin: '100px auto',
     '& > *': {
-      margin: theme.spacing(2, 0)
-    }
+      margin: theme.spacing(2, 0),
+    },
   },
   container: {
     '& > *': {
-      marginRight: theme.spacing(3)
-    }
+      marginRight: theme.spacing(3),
+    },
   },
   profile: {
     display: 'flex',
     marginRight: 'auto',
-    alignItems: 'center'
+    alignItems: 'center',
   },
   img: {
     width: '120px',
     height: '120px',
     borderRadius: '50%',
     backgroundColor: '#c4c4c4',
-    marginRight: theme.spacing(2)
+    marginRight: theme.spacing(2),
   },
   mapPin: {
     fontSize: '2rem',
     color: theme.palette.primary.main,
-    marginRight: theme.spacing(1)
+    marginRight: theme.spacing(1),
   },
   locContainer: {
     marginTop: theme.spacing(1),
     display: 'flex',
-    alignItems: 'center'
+    alignItems: 'center',
   },
   chipConatiner: {
     display: 'flex',
-    flexWrap: 'wrap'
+    flexWrap: 'wrap',
   },
   chips: {
     display: 'flex',
@@ -50,19 +52,19 @@ const useStyles = makeStyles(theme => ({
     marginRight: theme.spacing(2),
     marginBottom: theme.spacing(2),
     borderRadius: '4px',
-    backgroundColor: theme.colors.whiteGray
+    backgroundColor: theme.colors.whiteGray,
   },
   desc: {
-    color: theme.colors.gray
+    color: theme.colors.gray,
   },
   secondary: {
-    backgroundColor: theme.colors.secondaryLight
+    backgroundColor: theme.colors.secondaryLight,
   },
   bookmark: {
     color: theme.palette.primary.main,
     marginRight: theme.spacing(1),
-    transform: 'rotate(90deg)'
-  }
+    transform: 'rotate(90deg)',
+  },
 }));
 
 const Project = () => {
@@ -83,16 +85,16 @@ const Project = () => {
       'Registered Nurse',
       'Registered Nurse',
       'Registered Nurse',
-      'Registered Nurse'
-    ]
+      'Registered Nurse',
+    ],
   };
 
   return (
     <Grid container className={classes.root} xs={12} sm={8} direction="column">
       <Grid container className={classes.container} justify="flex-end" direction="row" alignItems="center">
         <div className={classes.profile}>
-          {/*// TODO :- remove the div and uncomment the img, when getting pic from backend API */}
-          <div className={classes.img}></div>
+          {/* // TODO :- remove the div and uncomment the img, when getting pic from backend API */}
+          <div className={classes.img} />
           {/* <img src="" alt="Profile Picture"/> */}
 
           <Grid item>
@@ -122,7 +124,7 @@ const Project = () => {
       </Typography>
       <Typography variant="h2">Looking for</Typography>
       <div className={classes.chipConatiner}>
-        {project.lookingFor.map(text => (
+        {project.lookingFor.map((text) => (
           <div className={classes.chips}>
             <Bookmark className={classes.bookmark} />
             <Typography variant="body2" className={classes.desc}>
