+++ frontend/src/pages/search.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 function SearchPage() {
   return <h1>Search Page</h1>;
