+++ frontend/src/pages/style.jsx
@@ -12,7 +12,7 @@ import FormHelperText from '@material-ui/core/FormHelperText';
 import Checkbox from '@material-ui/core/Checkbox';
 import { Divider, Button } from '@material-ui/core';
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles(() => ({
   root: {},
   sectionHeading: {
     textDecoration: 'underline',
