+++ frontend/src/styles/colors.js
@@ -28,7 +28,7 @@ const COLORS = {
   error: '#F44336',
   warning: '#FF9800',
   info: '#2196F3',
-  success: '#4CAF50'
+  success: '#4CAF50',
 };
 
 export default COLORS;
